refactor(agents): move pure pipeline helpers to backend-neutral candidates.py
Also update test patch targets to match the new module location.
