+++ test/unit/test_pipeline_agents.py
@@ -34,8 +34,10 @@ def test_extract_discovery_pairs_reads_books_from_dict_and_json():
 def test_extract_candidate_ids_calls_search(monkeypatch):
     state = {"targets": {"tropes": ["heist"], "styles": []}}
     with (
-        patch("agentic_librarian.agents.pipeline.search_internal_database", return_value=[{"id": "w1"}, {"id": "w2"}]),
-        patch("agentic_librarian.agents.pipeline.get_unacted_suggestions", return_value=[{"id": "w2"}]),
+        patch(
+            "agentic_librarian.agents.candidates.search_internal_database", return_value=[{"id": "w1"}, {"id": "w2"}]
+        ),
+        patch("agentic_librarian.agents.candidates.get_unacted_suggestions", return_value=[{"id": "w2"}]),
     ):
         ids = extract_candidate_ids(state)
     assert ids == ["w1", "w2"]  # de-duplicated, order preserved
